requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
blas=1.0=mkl
bottleneck=1.3.4=py39h080aedc_0
box2d-py=2.3.8=py39h415ef7b_5
brotli=1.0.9=ha925a31_2
ca-certificates=2022.4.26=haa95532_0
cairo=1.16.0=hb19e0ff_1008
certifi=2021.10.8=py39haa95532_2
cloudpickle=2.0.0=pyhd3eb1b0_0
cudatoolkit=11.3.1=h59b6b97_2
cycler=0.11.0=pyhd3eb1b0_0
expat=2.4.8=h39d44d4_0
ffmpeg=4.2.2=he774522_0
fontconfig=2.14.0=hce3cb01_0
fonttools=4.25.0=pyhd3eb1b0_0
freetype=2.10.4=hd328e21_0
fribidi=1.0.10=h62dcd97_0
future=0.18.2=py39haa95532_1
getopt-win32=0.1=h8ffe710_0
gettext=0.19.8.1=ha2e2712_1008
graphviz=2.47.1=h8277ce1_0
gts=0.7.6=h7c369d9_2
gym=0.21.0=py39h832f523_2
icu=68.1=h6c2663c_0
intel-openmp=2021.4.0=haa95532_3556
jpeg=9e=h2bbff1b_0
kiwisolver=1.3.2=py39hd77b12b_0
libclang=11.1.0=default_h5c34c98_1
libffi=3.4.2=h604cdb4_1
libgd=2.3.2=h138e682_0
libglib=2.68.4=h3be07f2_1
libiconv=1.16=h2bbff1b_2
libpng=1.6.37=h2a8f88b_0
libtiff=4.2.0=hd0e1b90_0
libuv=1.40.0=he774522_0
libwebp=1.2.2=h2bbff1b_0
libwebp-base=1.2.2=h2bbff1b_0
libxcb=1.13=hcd874cb_1004
lz4-c=1.9.3=h2bbff1b_1
m2w64-gcc-libgfortran=5.3.0=6
m2w64-gcc-libs=5.3.0=7
m2w64-gcc-libs-core=5.3.0=7
m2w64-gmp=6.1.0=2
m2w64-libwinpthread-git=5.0.0.4634.697f757=2
matplotlib=3.5.1=py39haa95532_1
matplotlib-base=3.5.1=py39hd77b12b_1
mkl=2021.4.0=haa95532_640
mkl-service=2.4.0=py39h2bbff1b_0
mkl_fft=1.3.1=py39h277e83a_0
mkl_random=1.2.2=py39hf11a4ad_0
msys2-conda-epoch=20160418=1
munkres=1.1.4=py_0
numexpr=2.8.1=py39hb80d3ca_0
numpy=1.21.5=py39ha4e8547_0
numpy-base=1.21.5=py39hc2deb75_0
openssl=1.1.1n=h2bbff1b_0
packaging=21.3=pyhd3eb1b0_0
pandas=1.4.1=py39hd77b12b_1
pango=1.42.4=hec34917_5
pcre=8.45=hd77b12b_0
pillow=9.0.1=py39hdc2b20a_0
pip=21.2.4=py39haa95532_0
pixman=0.40.0=h2bbff1b_1
pthread-stubs=0.3=h3c9f919_1
pyglet=1.5.16=py39hcbf5309_1
pyparsing=3.0.4=pyhd3eb1b0_0
pyqt=5.12.3=py39hcbf5309_8
pyqt-impl=5.12.3=py39h415ef7b_8
pyqt5-sip=4.19.18=py39h415ef7b_8
pyqtchart=5.12=py39h415ef7b_8
pyqtwebengine=5.12.1=py39h415ef7b_8
python=3.9.12=h6244533_0
python-dateutil=2.8.2=pyhd3eb1b0_0
python-graphviz=0.19.2=pypi_0
python_abi=3.9=2_cp39
pytorch=1.11.0=py3.9_cuda11.3_cudnn8_0
pytorch-mutex=1.0=cuda
pytz=2021.3=pyhd3eb1b0_0
qt=5.12.9=h5909a2a_4
setuptools=61.2.0=py39haa95532_0
six=1.16.0=pyhd3eb1b0_1
sqlite=3.38.2=h2bbff1b_0
tk=8.6.11=h2bbff1b_0
torchviz=0.0.2=pypi_0
tornado=6.1=py39h2bbff1b_0
typing_extensions=4.1.1=pyh06a4308_0
tzdata=2022a=hda174b7_0
vc=14.2=h21ff451_1
vs2015_runtime=14.27.29016=h5e58377_2
wheel=0.37.1=pyhd3eb1b0_0
wincertstore=0.2=py39haa95532_2
xorg-kbproto=1.0.7=hcd874cb_1002
xorg-libice=1.0.10=hcd874cb_0
xorg-libsm=1.2.3=hcd874cb_1000
xorg-libx11=1.7.2=hcd874cb_0
xorg-libxau=1.0.9=hcd874cb_0
xorg-libxdmcp=1.1.3=hcd874cb_0
xorg-libxext=1.3.4=hcd874cb_1
xorg-libxpm=3.5.13=hcd874cb_0
xorg-libxt=1.2.1=hcd874cb_2
xorg-xextproto=7.3.0=hcd874cb_1002
xorg-xproto=7.0.31=hcd874cb_1007
xz=5.2.5=h8cc25b3_1
zlib=1.2.12=h8cc25b3_2
zstd=1.4.9=h19a0ad4_0
